[ui] Handle conditional filter in postgrest query builder
