+++ app/src/app/search/hooks/use-search.ts
@@ -1,17 +1,16 @@
 import useSWR, {Fetcher} from "swr";
 import type {SearchFilter, SearchResult} from "@/app/search/search.types";
-import {generateFTSQuery} from "@/app/search/hooks/use-filter";
 
 const fetcher: Fetcher<SearchResult, string> = (...args) => fetch(...args).then(res => res.json())
 
 export default function useSearch(filters: SearchFilter[]) {
-    const searchParams = filters
-        .map(f => [f.name, generatePostgrestQuery(f)])
-        .filter(([, query]) => !!query)
-        .reduce((params, [name, query]) => {
-            params.set(name!, query!);
-            return params;
-        }, new URLSearchParams());
+  const searchParams = filters
+    .map(f => [f.name, generatePostgrestQuery(f)])
+    .filter(([, query]) => !!query)
+    .reduce((params, [name, query]) => {
+      params.set(name!, query!);
+      return params;
+    }, new URLSearchParams());
 
   const search = useSWR<SearchResult>(`/search/api?${searchParams}`, fetcher, {
     keepPreviousData: true,
@@ -21,8 +20,22 @@ export default function useSearch(filters: SearchFilter[]) {
   return {search, searchParams}
 }
 
+export function generateFTSQuery(prompt: string = ""): string | null {
+  const cleanedPrompt = prompt.trim().toLowerCase();
+  const isNegated = (word: string) => new RegExp(`\\-\\b(${word})\\b`).test(cleanedPrompt)
+  const uniqueWordsInPrompt = new Set(cleanedPrompt.match(/\b\w+\b/g) ?? []);
+  const tsQuery = [...uniqueWordsInPrompt]
+    .map(word => isNegated(word) ? `!'${word}':*` : `'${word}':*`)
+    .join(' & ');
+
+  return tsQuery ? `fts(simple).${tsQuery}` : null;
+}
+
 const generatePostgrestQuery = (f: SearchFilter) => {
-  // TODO: generate type safe filter name type to ensure all branches are covered
+  if (f.type === 'conditional') {
+    return f.condition && f.selected.length ? `${f.condition}.${f.selected[0]}` : null
+  }
+
   switch (f.name) {
     case 'search':
       return generateFTSQuery(f.selected[0])
@@ -34,8 +47,6 @@ const generatePostgrestQuery = (f: SearchFilter) => {
       return f.selected.length ? `cd.${f.selected.join()}` : null
     case 'primary_activity_category_path':
       return f.selected.length ? `cd.${f.selected.join()}` : null
-    case 'conditional':
-      return f.condition && f.selected.length ? `${f.condition}.${f.selected[0]}` : null
     default:
       return null
   }
