+++ app/src/app/search/search.types.ts
@@ -2,38 +2,46 @@ import {Tables} from "@/lib/database.types";
 
 export type SearchFilterCondition = "eq" | "gt" | "lt" | "in" | "ilike"
 
+export type SearchFilterName =
+  "search"
+  | "tax_reg_ident"
+  | "unit_type"
+  | "physical_region_path"
+  | "primary_activity_category_path"
+  | string
+
 export type SearchFilterOption = {
-    readonly label: string
-    readonly value: string
-    readonly humanReadableValue?: string
-    readonly className?: string
+  readonly label: string
+  readonly value: string
+  readonly humanReadableValue?: string
+  readonly className?: string
 }
 
 export type SearchFilter = {
-    readonly type: "options" | "radio" | "conditional" | "search"
-    readonly name: string
-    readonly label: string
-    readonly options?: SearchFilterOption[]
-    readonly selected: string[]
-    readonly condition?: SearchFilterCondition
+  readonly type: "options" | "radio" | "conditional" | "search"
+  readonly name: SearchFilterName
+  readonly label: string
+  readonly options?: SearchFilterOption[]
+  readonly selected: string[]
+  readonly condition?: SearchFilterCondition
 }
 
 export type SearchResult = {
-    statisticalUnits: Partial<Tables<"statistical_unit">>[]
-    count: number
+  statisticalUnits: Partial<Tables<"statistical_unit">>[]
+  count: number
 }
 
 export interface ConditionalValue {
-    condition: SearchFilterCondition
-    value: string,
+  condition: SearchFilterCondition
+  value: string,
 }
 
 interface ToggleOption {
-    type: "toggle_option",
-    payload: {
-        name: string,
-        value: string
-    }
+  type: "toggle_option",
+  payload: {
+    name: string,
+    value: string
+  }
 }
 
 interface ToggleRadioOption {
@@ -45,12 +53,12 @@ interface ToggleRadioOption {
 }
 
 interface SetCondition {
-    type: "set_condition",
-    payload: {
-        name: string,
-        value: string,
-        condition: SearchFilterCondition
-    }
+  type: "set_condition",
+  payload: {
+    name: string,
+    value: string,
+    condition: SearchFilterCondition
+  }
 }
 
 interface SetSearch {
@@ -62,14 +70,14 @@ interface SetSearch {
 }
 
 interface Reset {
-    type: "reset",
-    payload: {
-        name: string
-    }
+  type: "reset",
+  payload: {
+    name: string
+  }
 }
 
 interface ResetAll {
-    type: "reset_all"
+  type: "reset_all"
 }
 
 export type SearchFilterActions = ToggleOption | ToggleRadioOption | SetCondition | SetSearch | Reset | ResetAll
